requirements.txt
beautifulsoup4==4.12.3
docx==0.2.4
dotenv-python==0.0.1
langchain==0.2.3
langchain-community==0.0.37
langchain-core==0.2.5
langchain-openai==0.1.8
langchainhub==0.1.15
openai==1.26.0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.31.0